+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Properties are now emitted as `readonly` when applicable.
 
 ## [0.3.4] - 2017-12-20
 - Handle optional and rest parameters in function type expressions.